Extract Task Review logic into separate object
