+++ src/Components/TaskReview/TaskReview.js
@@ -1,57 +1,18 @@
 import React from 'react';
 import TaskReviewStats from './TaskReviewStats.js';
+import { TaskReviewService } from './TaskReviewLogic';
 import QAScreen from './QAScreen.js'
 import QuestionReview from './QuestionReview'
 import { useHistory } from 'react-router';
 import { useState } from 'react';
 import './TaskReview.css';
 
-function getTaskQuestions(task){
-  let questions = [];
-  task.pages.forEach(page => [
-    page.blocks.forEach(block => {
-      if(block.__typename == 'QuizBlock'){
-        questions = questions.concat(block.questions);
-      }
-    })
-  ])
-  return questions;
-}
-
-function getAnswerForQA(qa, questions){
-  const question = questions.filter(question => question.id == qa.question.id)[0]
-
-  if(!question || question.__typename == 'FrQuestion'){
-    return qa.answer.answer;
-  }
-
-  const index = parseInt(qa.answer.answer);
-
-  return {
-    answer: question.options[index].description,
-    correctAnswer:  question.options[qa.question.answers[0]].description
-  };
-}
-
-function getTaskQAs(task, qas){
-  const questions = getTaskQuestions(task);
-  return qas.map(qa => {
-    return {
-      ...qa,
-      answer: {
-        ...(qa.answer),
-        ...(getAnswerForQA(qa, questions))
-      }
-    }
-  })
-}
-
 export default function TaskReview(props) {
 
   const submission = props?.location?.state?.submitTask;
   const task = props?.location?.state?.task;
   const hist = useHistory();
-  const taskQAs = getTaskQAs(task, submission.questionAndAnswers);
+  const taskQAs = TaskReviewService.addAnswerValuesToQAs(task, submission.questionAndAnswers);
   const [focusedQA, setFocusedQA] = useState(taskQAs[0]);
 
   const mapRequirements = (req) => {
